+++ doc/CHANGELOG.rst
@@ -39,6 +39,7 @@ Version 5.0.0:
    interface-address -> interface-addresses, neighbor-address -> neighbor-addresses
  * Compatibility: change JSON for sr_capability_flags to be sr-capability-flags and data format
  * Compatibility: change node-descriptors to be list
+ * Compatibility: remove L from target in JSON extended communities
 
 Version 4.2.7:
  * Feature: logging parsing in debug mode will now print the JSON of updates